requirements.txt
# STマイクロRAGシステム 依存関係
streamlit>=1.46.0
openai>=1.90.0
langchain>=0.2.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0